Rename NoteProperty to LazyBindable
